Revert "Move PixelThresholdClusterizer (back?) to interface+src in order to use it outside of RecoLocalTracker/SiPixelClusterizer"
This reverts commit 766e9676854a1b785a1afe6c5800c7f3f4ec917f.
