+++ RecoLocalTracker/SiPixelClusterizer/BuildFile.xml
@@ -1,8 +1,4 @@
-<use   name="FWCore/ParameterSet"/>
 <use   name="FWCore/Utilities"/>
-<use   name="DataFormats/Common"/>
-<use   name="DataFormats/SiPixelDetId"/>
-<use   name="DataFormats/SiPixelCluster"/>
 <use   name="CalibTracker/SiPixelESProducers"/>
 <use   name="HeterogeneousCore/CUDACore"/>
 <use   name="cuda"/>
