+++ RecoLocalTracker/SiPixelClusterizer/plugins/BuildFile.xml
@@ -1,3 +1,7 @@
+<use   name="DataFormats/Common"/>
+<use   name="FWCore/ParameterSet"/>
+<use   name="DataFormats/SiPixelDetId"/>
+<use   name="DataFormats/SiPixelCluster"/>
 <use   name="boost_serialization"/>
 <use   name="RecoLocalTracker/SiPixelClusterizer"/>
 <use   name="RecoTracker/Record"/>
